Add `Profile` to public API; rename `profile` module to `profile_` to prevent collision with Python builtin module
