+++ spond_classes/member.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 from pydantic import BaseModel, Field
 
-from .profile import Profile
+from .profile_ import Profile
 
 
 class Member(BaseModel):
